76 - 05 - Enhancement- Protecting model relations

+++ app/Http/Controllers/Backend/BrandController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\backend;
 use App\DataTables\BrandDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Brand;
+use App\Models\Product;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
 use Str;
@@ -101,6 +102,9 @@ class BrandController extends Controller
     public function destroy(string $id)
     {
         $brand = Brand::findOrFail($id);
+        if (Product::where('brand_id', $brand->id)->count() > 0) {
+            return response(['status' => 'error', 'message' => 'This brand have products you can\'t delete it']);
+        }
         $brand->delete();
         $this->deleteImage($brand->logo);
